fix: 处理上传文件大小限制不准确的问题

+++ server/internal/domain/dto/config.go
@@ -10,8 +10,8 @@ type EmailConfigReq struct {
 }
 
 type StorageConfigReq struct {
-	MaxImgSize   int
-	MaxVideoSize int
+	MaxImgSize   int64
+	MaxVideoSize int64
 
 	Type      string
 	KeyID     string
